src/main/java/com/services/pm/product/service/ProductService.java
package com.services.pm.product.service;

import com.services.pm.product.model.Product;
import com.services.pm.product.repository.ProductRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
public class ProductService implements IProductService {

    @Autowired
    ProductRepository productRepository;

    @Override
    public List<Product> findAll() {
        return productRepository.findAll();
    }

    @Override
    public Product findById(Long id) {
        //Optional<Product> product = productRepository.findById(id);
        //return product.isPresent()? product.get() : null;
        return productRepository.getOne(id);
    }

    @Override
    public void deleteById(Long id) {
        productRepository.deleteById(id);
    }

    @Override
    public Product save(Product product) {
        return productRepository.save(product);
    }
}
